refactor(error): group the domain catalogs under catalog/
The shared/error/ root mixed two kinds of file: the machinery (registry,
domainError, index, boundaryLog, validation) and the eight per-domain
catalogs, which were 8 of 13 files and all the same shape repeated. Move
the catalogs into shared/error/catalog/ so the root holds only the
mechanism and the subfolder holds the data.

Pure move: each catalog's sole import (`./registry` -> `../registry`)
and the aggregator's imports in index.ts are the only line changes. No
one imports a catalog directly — only index.ts aggregates them — so the
blast radius is just those two spots. validation.ts stays at the root
(it's client-side validation messages, not a domain catalog).

Docs realigned: ach.md tree + §3.2, and rule 33's catalog path.

tsc clean, suite 395/395, build succeeds.
