+++ docs/.afm-log/events/2026-08.md
@@ -68,3 +68,7 @@
 - [2026-08-22T19:47:08Z] doc | s=5d247c | edit em docs/rubrics/error-classification.md
 - [2026-08-22T19:49:05Z] feature-spec | s=5d247c | edit em docs/features/024-error-boundary-centralization/spec.md
 - [2026-08-22T19:49:33Z] task | s=5d247c | commit
+- [2026-08-23T05:37:34Z] task | s=5d247c | commit
+- [2026-08-23T05:39:58Z] ach | s=5d247c | edit em docs/ach.md
+- [2026-08-23T05:40:03Z] ach | s=5d247c | edit em docs/ach.md
+- [2026-08-23T05:40:09Z] rule | s=5d247c | edit em docs/afm.md
