+++ docs/afm.md
@@ -182,7 +182,7 @@ Toda regra abaixo tem **gatilho executável** que o agente roda no teclado — p
 32. **Nenhum commit direto em `main`/`develop`.** Todo trabalho de código acontece num branch de feature criado a partir de `develop` (nome em inglês, ex. `feature/016-search-content`), commits seguindo `.commitlintrc` (Conventional Commits, em inglês). PR contra `develop` usando `.github/pull_request_template.md`; só o dono aprova/faz merge no GitHub. `main` só recebe merge de `develop` em release/deploy, nunca commit ou merge de feature branch direto. Branch protection no GitHub (`main`/`develop`, PR obrigatório) é camada extra — o enforcement primário é este check antes de commitar.
     *Verificação:* `git rev-parse --abbrev-ref HEAD` não é `main` nem `develop`. Se for, cria/troca pra branch de feature antes de qualquer `git commit`.
 
-33. **Erro no boundary é classificado: recuperável (`DomainError`) vs. bug.** Falha esperada de negócio é um `DomainError` (lançado no domain, traduzido no boundary — regra 15); throw inesperado é bug. O boundary **distingue os dois** — traduz `DomainError` → `TRPCError` e **relança o resto** — e nunca dressa um bug como erro de domínio recuperável. O logging central (`onError` do fetch adapter + `src/server/caller.ts`, via `logBoundaryError`) roteia por `level`: bug → `error` com stack; recuperável → seu próprio `level`. Metadata (`retryable`/`level`, `ErrorLevel = fatal|error|warn|info`) vive no catálogo `src/shared/error/*.ts` e é resolvida pelo `DomainError` com defaults `retryable=false`/`level=warn`. Ver ADR-0018 (estende ADR-0017; `Result<T,E>` foi avaliado e rejeitado).
+33. **Erro no boundary é classificado: recuperável (`DomainError`) vs. bug.** Falha esperada de negócio é um `DomainError` (lançado no domain, traduzido no boundary — regra 15); throw inesperado é bug. O boundary **distingue os dois** — traduz `DomainError` → `TRPCError` e **relança o resto** — e nunca dressa um bug como erro de domínio recuperável. O logging central (`onError` do fetch adapter + `src/server/caller.ts`, via `logBoundaryError`) roteia por `level`: bug → `error` com stack; recuperável → seu próprio `level`. Metadata (`retryable`/`level`, `ErrorLevel = fatal|error|warn|info`) vive nos catálogos `src/shared/error/catalog/*.ts` e é resolvida pelo `DomainError` com defaults `retryable=false`/`level=warn`. Ver ADR-0018 (estende ADR-0017; `Result<T,E>` foi avaliado e rejeitado).
     *Verificação:* `rg -l "new TRPCError" src/server/features/*/procedures/*.ts` retorna vazio — a tradução não mora mais na procedure, e sim no único choke point `src/server/http/domainErrorToTRPCError.ts`, que devolve `null` pra throw não-`DomainError` (bug segue bug, nunca é embrulhado como erro de domínio). **Compliant hoje** — centralizada em `024-error-boundary-centralization` (2026-08-22, ADR-0019).
     *Nota de 2026-08-22:* o gatilho anterior (`comm -23` entre quem constrói `TRPCError` e quem ramifica em `instanceof DomainError`) media a disciplina **dentro** de cada procedure. Com a tradução centralizada, o primeiro conjunto passou a ser vazio e o `comm` passaria **vacuously** — verificando nada. O gatilho novo mira onde a convenção agora vive.
